+++ inventarios/modulos/extintores/recarga/ajax-grid-data.php
@@ -85,7 +85,7 @@ while( $row=mysqli_fetch_array($query) ) {  // preparing an array
     $nestedData[] = date("d/m/Y", strtotime($row["FechaDeRegistro"]));
     $nestedData[] = '<td><center>
                     <a href="editar.php?ExtintorID='.$row['ExtintorID'].'"  data-toggle="tooltip" title="Editar datos" class="btn btn-sm btn-info"> <i class="fa-solid fa-pen-to-square" style="color: #f2eded;"></i> </a>
-                    <a href="index.php?action=delete&ExtintorID='.$row['ExtintorID'].'"  data-toggle="tooltip" title="Eliminar" class="btn btn-sm btn-danger"> <i class="fa-solid fa-trash-can" style="color: #f2eded;"></i> </a>
+                    <a href="index.php?action=delete&ExtintorID='.$row['ExtintorID'].'" onclick="return confirmacion()" data-toggle="tooltip" title="Eliminar" class="btn btn-sm btn-danger"> <i class="fa-solid fa-trash-can" style="color: #f2eded;"></i> </a>
                     </center></td>';		
 	
 	$data[] = $nestedData;
